Appease linter

+++ contracts/v6/AaveLPManager.sol
@@ -1,29 +1,50 @@
 // SPDX-License-Identifier: agpl-3.0
 pragma solidity ^0.6.0;
 
-import { Context } from "@openzeppelin/contracts-v6/utils/Context.sol";
-import { LendingPool } from "@aave/protocol-v2/contracts/protocol/lendingpool/LendingPool.sol";
-import { LendingPoolAddressesProvider } from "@aave/protocol-v2/contracts/protocol/configuration/LendingPoolAddressesProvider.sol";
-import { IERC20 } from "@openzeppelin/contracts-v6/token/ERC20/IERC20.sol";
+import {Context} from "@openzeppelin/contracts-v6/utils/Context.sol";
+import {LendingPool} from "@aave/protocol-v2/contracts/protocol/lendingpool/LendingPool.sol";
+import {LendingPoolAddressesProvider} from "@aave/protocol-v2/contracts/protocol/configuration/LendingPoolAddressesProvider.sol";
+import {IERC20} from "@openzeppelin/contracts-v6/token/ERC20/IERC20.sol";
 
-import { AaveLPManagerLib } from "./AaveLPManagerLib.sol";
+import {AaveLPManagerLib} from "./AaveLPManagerLib.sol";
 
 contract AaveLPManager is Context {
     modifier aaveInitLock() {
-        require(isAaveInit(), string(abi.encodePacked(AaveLPManagerLib.NAMESPACE, ": ", "UNINITIALIZED")));
+        require(
+            isAaveInit(),
+            string(
+                abi.encodePacked(
+                    AaveLPManagerLib.NAMESPACE,
+                    ": ",
+                    "UNINITIALIZED"
+                )
+            )
+        );
         _;
     }
 
     //Basically a constructor, but the hardhat-deploy plugin does not support diamond contracts with facets that has
     // constructors. We imitate a constructor with a one-time only function. This is called immediately after deployment
-    function initAaveLPManager(address _lendingPoolProviderAddr, address[] memory _supportedTokens) public {
-        require(!isAaveInit(), string(abi.encodePacked(AaveLPManagerLib.NAMESPACE, ": ", "ALREADY_INITIALIZED")));
+    function initAaveLPManager(
+        address _lendingPoolProviderAddr,
+        address[] memory _supportedTokens
+    ) public {
+        require(
+            !isAaveInit(),
+            string(
+                abi.encodePacked(
+                    AaveLPManagerLib.NAMESPACE,
+                    ": ",
+                    "ALREADY_INITIALIZED"
+                )
+            )
+        );
 
         AaveLPManagerLib.StateStorage storage ss = AaveLPManagerLib.getState();
 
         ss.lpProviderAddr = _lendingPoolProviderAddr;
 
-        for (uint i = 0; i < _supportedTokens.length; i++) {
+        for (uint256 i = 0; i < _supportedTokens.length; i++) {
             addAaveSupportedToken(_supportedTokens[i]);
         }
 
@@ -41,15 +62,37 @@ contract AaveLPManager is Context {
     }
 
     //TODO: Need to make sure this is locked down to only owner and approved callers (eg chainlink)
-    function depositToAave(address erc20TokenAddr, uint256 amount) public aaveInitLock {
-        (bool _isSupportedToken,) = isAaveSupportedToken(erc20TokenAddr);
-        require(_isSupportedToken, string(abi.encodePacked(AaveLPManagerLib.NAMESPACE, ": ", "This token is not currently supported")));
+    function depositToAave(address erc20TokenAddr, uint256 amount)
+        public
+        aaveInitLock
+    {
+        (bool _isSupportedToken, ) = isAaveSupportedToken(erc20TokenAddr);
+        require(
+            _isSupportedToken,
+            string(
+                abi.encodePacked(
+                    AaveLPManagerLib.NAMESPACE,
+                    ": ",
+                    "This token is not currently supported"
+                )
+            )
+        );
 
         IERC20(erc20TokenAddr).approve(getAaveLPAddr(), amount);
-        LendingPool(getAaveLPAddr()).deposit(erc20TokenAddr, amount, address(this), 0);
+        LendingPool(getAaveLPAddr()).deposit(
+            erc20TokenAddr,
+            amount,
+            address(this),
+            0
+        );
     }
 
-    function isAaveSupportedToken(address tokenAddr) public view aaveInitLock returns (bool, uint256) {
+    function isAaveSupportedToken(address tokenAddr)
+        public
+        view
+        aaveInitLock
+        returns (bool, uint256)
+    {
         AaveLPManagerLib.StateStorage storage ss = AaveLPManagerLib.getState();
 
         bool _isSupportedToken = false;
@@ -71,8 +114,7 @@ contract AaveLPManager is Context {
         //Check to ensure that the token is in fact an ERC-20 token. This isn't full proof, but there is no efficient
         // way to definitively check a contract implements all ERC-20 functions. Since balanceOf() is a view function,
         // we use that to efficiently check if it is an ERC-20 token
-        try IERC20(tokenAddr).balanceOf(address(this)) {
-        } catch {
+        try IERC20(tokenAddr).balanceOf(address(this)) {} catch {
             revert("The tokenAddr supplied is not ERC20 compatible");
         }
 
@@ -82,25 +124,43 @@ contract AaveLPManager is Context {
     function removeAaveSupportedToken(address tokenAddr) public aaveInitLock {
         AaveLPManagerLib.StateStorage storage ss = AaveLPManagerLib.getState();
 
-        (bool _isSupportedToken, uint256 _tokenIndex) = isAaveSupportedToken(tokenAddr);
+        (bool _isSupportedToken, uint256 _tokenIndex) = isAaveSupportedToken(
+            tokenAddr
+        );
         if (_isSupportedToken) {
             //Remove the token without preserving order
-            ss.supportedTokens[_tokenIndex] = ss.supportedTokens[ss.supportedTokens.length - 1];
+            ss.supportedTokens[_tokenIndex] = ss.supportedTokens[
+                ss.supportedTokens.length - 1
+            ];
             ss.supportedTokens.pop();
         }
     }
 
-    function getAllAaveSupportedTokens() public view aaveInitLock returns (address[] memory) {
+    function getAllAaveSupportedTokens()
+        public
+        view
+        aaveInitLock
+        returns (address[] memory)
+    {
         AaveLPManagerLib.StateStorage storage ss = AaveLPManagerLib.getState();
         return ss.supportedTokens;
     }
 
-    function updateAaveTokenTargetedPercentage(address tokenAddr, uint targetPercentage) public aaveInitLock {
+    function updateAaveTokenTargetedPercentage(
+        address tokenAddr,
+        uint256 targetPercentage
+    ) public aaveInitLock {
         AaveLPManagerLib.StateStorage storage ss = AaveLPManagerLib.getState();
         ss.tokenMetadata[tokenAddr].targetPercent = targetPercentage;
     }
 
-    function getAaveIncome(address tokenAddr) public view aaveInitLock returns (uint256) {
-        return LendingPool(getAaveLPAddr()).getReserveNormalizedIncome(tokenAddr);
+    function getAaveIncome(address tokenAddr)
+        public
+        view
+        aaveInitLock
+        returns (uint256)
+    {
+        return
+            LendingPool(getAaveLPAddr()).getReserveNormalizedIncome(tokenAddr);
     }
 }
