+++ contracts/v6/AaveLPManagerLib.sol
@@ -9,14 +9,14 @@ library AaveLPManagerLib {
     struct StateStorage {
         bool isInit;
         address[] supportedTokens;
-        mapping (address => TokenMetadata) tokenMetadata;
+        mapping(address => TokenMetadata) tokenMetadata;
         address lpProviderAddr;
         address lpAddr;
     }
 
     struct TokenMetadata {
-        uint targetPercent;
-        uint actualPercent;
+        uint256 targetPercent;
+        uint256 actualPercent;
     }
 
     function getState()
